# api/swagger.yml
swagger: "2.0"
info:
  description: "Squabble API"
  version: "0.0.1"
  title: "Squabble"
#  termsOfService: ""
#  contact:
#    email: "test@mail.de"
#  license:
#    name: "Apache 2.0"
#    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "squabble.com"
basePath: "/"
...
schemes:
- "http"
paths:
  /accounts:
    post:
      tags:
      - "accounts"
      summary: "Erstellt einen Account."
#      description: ""
#      operationId: "addAccount"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Account-Objekt, das erstellt werden soll."
        required: true
        schema:
          $ref: "#/definitions/RegisterAccountData"
      responses:
        "201":
          description: "Benutzer wurde erfolgreich erstellt."
          schema:
            $ref: "#/definitions/ApiSuccessResponseSimple"
        "409":
          description: "Email oder Benutzername ist vergeben."
          schema:
            $ref: "#/definitions/ApiErrorResponse"
        "422":
          description: "Falsche Eingabedaten. Z. B. unzulässiger Benutzername, Passwort oder Tierart."
          schema:
            $ref: "#/definitions/ApiErrorResponse"
#      security:
#      - petstore_auth:
#        - "write:pets"
#        - "read:pets"
  /accounts/{accountName}:
    get:
      tags:
      - "accounts"
      summary: "Gibt Informationen zu einem Account zurück."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "accountName"
        in: "path"
        description: "Benutzername des Accounts, von dem Informationen zurückgegeben werden sollen."
        required: true
        type: "string"
      responses:
        "200":
          description: "Accountinformationen wurden erfolgreich zurückgegeben."
          schema:
            type: "object"
            properties:
              code:
                type: "string"
                example: "SUCCESS"
              devMsg:
                type: "string"
                example: ""
              content:
                $ref: "#/definitions/AccountInfo"
        "404":
          description: "Account wurde nicht gefunden."
          schema:
            $ref: "#/definitions/ApiErrorResponse"
    patch:
      tags:
      - "accounts"
      summary: "Aktualisiert einen vorhandenen Account."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "accountName"
        in: "path"
        description: "Benutzername des Accounts, der aktualisiert werden soll. Attribute, die nicht aktualisiert werden sollen, werden weggelassen."
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Account-Objekt mit Attributen, die aktualisiert werden sollen."
        required: true
        schema:
          $ref: "#/definitions/UpdateAccountData"
      responses:
        "200":
          description: "Benutzer wurde erfolgreich aktualisiert."
          schema:
            $ref: "#/definitions/ApiSuccessResponseSimple"
        "404":
          description: "Account wurde nicht gefunden."
          schema:
            $ref: "#/definitions/ApiErrorResponse"
        "409":
          description: "Email oder Benutzername ist vergeben."
          schema:
            $ref: "#/definitions/ApiErrorResponse"
        "422":
          description: "Falsche Eingabedaten. Z. B. unzulässiger Benutzername, Passwort oder Tierart."
          schema:
            $ref: "#/definitions/ApiErrorResponse"
    delete:
      tags:
      - "accounts"
      summary: "Löscht einen vorhandenen Account."
      produces:
      - "application/json"
      parameters:
      - name: "accountName"
        in: "path"
        description: "Benutzername des Accounts, der gelöscht werden soll."
        required: true
        type: "string"
      responses:
        "200":
          description: "Benutzer wurde erfolgreich gelöscht."
          schema:
            $ref: "#/definitions/ApiSuccessResponseSimple"
        "404":
          description: "Account wurde nicht gefunden."
          schema:
            $ref: "#/definitions/ApiErrorResponse"
...
  /accounts/{accountName}/friendrequests/create:
    post:
      tags:
      - "friends"
      summary: "Sendet dem angegebenen Benutzer eine Freundschaftsanfrage."
      produces:
      - "application/json"
      parameters:
      - name: "accountName"
        in: "path"
        description: "Name des Accounts, der eine Freundschaftsanfrage versenden soll."
        required: true
        type: "string"
      - name: "friendName"
        in: "query"
        description: "Name des Accounts, der eine Freundschaftsanfrage erhalten soll."
        required: true
        type: "string"
      responses:
        "201":
          description: "Die Freundschaftsanfrage wurde erfolgreich versendet."
          schema:
            $ref: "#/definitions/ApiSuccessResponseSimple"
        "404":
          description: "Eines der beiden Accounts wurde nicht gefunden."
          schema:
            $ref: "#/definitions/ApiErrorResponse"
        "409":
          description: "Der Account ist schon mit dem angegebenen Freund befreundet oder es wurde schon eine Freundschaftsanfrage versendet oder es wurde eine Freundschaftsanfrage an sich selbst versendet."
          schema:
            $ref: "#/definitions/ApiErrorResponse"
  /accounts/{accountName}/friendrequests/accept:
    post:
      tags:
      - "friends"
      summary: "Akzeptiert die Freundschaftsanfrage des angegebenen Benutzers."
      produces:
      - "application/json"
      parameters:
      - name: "accountName"
        in: "path"
        description: "Name des Accounts, der eine Freundschaftsanfrage akzeptieren soll."
        required: true
        type: "string"
      - name: "friendName"
        in: "query"
        description: "Name des Accounts, von dem die Freundschaftsanfrage akzeptiert werden soll."
        required: true
        type: "string"
      responses:
        "200":
          description: "Die Freundschaftsanfrage wurde erfolgreich akzeptiert."
          schema:
            $ref: "#/definitions/ApiSuccessResponseSimple"
        "404":
          description: "Eines der beiden Accounts wurde nicht gefunden."
          schema:
            $ref: "#/definitions/ApiErrorResponse"
        "409":
          description: "Es gibt keine passende Freundschaftsanfrage."
          schema:
            $ref: "#/definitions/ApiErrorResponse"
  /accounts/{accountName}/friendrequests/decline:
    post:
      tags:
      - "friends"
      summary: "Lehnt die Freundschaftsanfrage des angegebenen Benutzers ab."
      produces:
      - "application/json"
      parameters:
      - name: "accountName"
        in: "path"
        description: "Name des Accounts, der eine Freundschaftsanfrage ablehnen soll."
        required: true
        type: "string"
      - name: "friendName"
        in: "query"
        description: "Name des Accounts, von dem die Freundschaftsanfrage abgelehnt werden soll."
        required: true
        type: "string"
      responses:
        "200":
          description: "Die Freundschaftsanfrage wurde erfolgreich abgelehnt."
          schema:
            $ref: "#/definitions/ApiSuccessResponseSimple"
        "404":
          description: "Eines der beiden Accounts wurde nicht gefunden."
          schema:
            $ref: "#/definitions/ApiErrorResponse"
        "409":
          description: "Es gibt keine passende Freundschaftsanfrage."
          schema:
            $ref: "#/definitions/ApiErrorResponse"
...
  /accounts/login:
    post:
      tags:
      - "accounts"
      summary: "Loggt einen Account ein."
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "query"
        description: "Der Benutzername."
        required: true
        type: "string"
      - name: "password"
        in: "query"
        description: "Das Passwort."
        required: true
        type: "string"
      responses:
        "200":
          description: "Der Benutzer wurde erfolgreich eingeloggt."
          schema:
            $ref: "#/definitions/ApiSuccessResponseSimple"
#          headers:
#            X-Rate-Limit:
#              type: "integer"
#              format: "int32"
#              description: "calls per hour allowed by the user"
#            X-Expires-After:
#              type: "string"
#              format: "date-time"
#              description: "date in UTC when token expires"
        "409":
          description: "Falsche Email oder falsches Passwort."
          schema:
            $ref: "#/definitions/ApiErrorResponse"
...
definitions:
  AccountInfo:
    type: "object"
    properties:
      email:
        type: "string"
        example: "my_epic_mail@gmail.com"
        description: "Registrierungs-Email"
      username:
        type: "string"
        example: "EpicRobloxMaster"
        description: "Einzigartiger Name"
      pictureId:
        type: "integer"
        description: "ID des ausgewählten Profilicons"
      petId:
        type: "integer"
        description: "ID des PokePet"
  RegisterAccountData:
    type: "object"
    required:
    - "email"
    - "username"
    - "password"
    - "petName"
    - "petBreed"
    properties:
      email:
        type: "string"
        example: "my_epic_mail@gmail.com"
        description: "Registrierungs-Email"
      username:
        type: "string"
        example: "EpicRobloxMaster"
        description: "Einzigartiger Name"
      password:
        type: "string"
        example: "password123"
        description: "Klartext Passwort"
      pictureId:
        type: "integer"
        description: "ID des ausgewählten Profilicons"
      petName:
        type: "string"
        example: "Sparky"
        description: "Name des PokeTier"
      petKind:
        type: "string"
        example: "KANGAROO"
        description: "Art des Poketier"
      petImage:
        type: "string"
        example: "/9j/4AAQSkZJRgABAQAAAQABAAD/2wBDABQODxIPDRQSEBIXFRQYHjIhHhwcHj0sLiQySUBMS0dARkVQWnNiUFVtVkVGZIhlbXd7gYKBTmCNl4x9lnN+gXz/2wBDARUXFx4aHjshITt8U0ZTfHx8fHx8fHx8fHx8fHx8fHx8fHx8fHx8fHx8fHx8fHx8fHx8fHx8fHx8fHx8fHx8fHz/wAARCAA0AEYDASIAAhEBAxEB/8QAGgAAAgMBAQAAAAAAAAAAAAAAAAQDBQYBAv/EAC0QAAEDAwMCBQMFAQAAAAAAAAEAAgMEESEFEjETQQYUIlFhMnGRI4GhscHR/8QAFwEAAwEAAAAAAAAAAAAAAAAAAAECA//EABkRAQEBAQEBAAAAAAAAAAAAAAABETECEv/aAAwDAQACEQMRAD8AT12TdVW9gqcp/VX76uT4wkVHrrSPBQulccCGhxBseDbBSNzcrOfSyzTDIbiYDcc/wodFgE+oxl7SY4zud7fCv3vNS50TGE7j3wLI4XWJIQrqr8O1cDdzbPb7DlcodFL5L1ILWt5bwSr2I+apgCeEJvUKXylSY+WHLfshMGKyTfUSO93FQXTQoJZHXc4NB7lejpp4a+5+As9jTKNLphU1DeoAY72ye60DqJsMXQeyN9M44BGAldHpDSxSNqItwc7FwraH1EMiLy3u1+R+VN628ZI8aVQ08NM8BtmAnnuPlIV2+nmM8FQGxBvqYWbvwndSqxTERbrF3+JTfEaZwPqLhk8q5GFqSPxAZHCN9K120BxAdY2TM80NZFup2kOHIIsR8EKgZGGSnptsTguAz8K/8O0klJSydc36hG25+/8A1OxMrJa/v88N7C1oaA2/dCtfFksBmjhDbyNySOwQnCqelpuu67+FbQU0UQAA572XiKBrOLj90w1pHOVEmLvrRIQ1nO0fZRMd+phyY+rGFxkZvnITpSqXW6Gaoe2dgJDebdlDTUfVp2mMESD6hdaFzbNe12WPGfhVNVSCkjkm8x02NFxblEBKaJ9C9skwsx3seMo1LXNsTYKR/rH1PA4VDVVktS473OIv3N1E1Um1MXl5LnG7jkk90LgtblCZN3fBK9E4BQhSpLGAeylshCQRycFZ7xC4iheL4uP7QhEFZReghCtKRCEID//Z"
        description: "Optionales, in Base64 codiertes, Bild des PokeTier. Max. ca. 4 KB."
  UpdateAccountData:
    type: "object"
    properties:
      email:
        type: "string"
        example: "my_new_mail@gmail.com"
        description: "Registrierungs-Email"
      username:
        type: "string"
        example: "MyBetterName"
        description: "Einzigartiger Name"
      password:
        type: "string"
        example: "myBetterPassword123"
        description: "Klartext Passwort"
      pictureId:
        type: "integer"
        description: "ID des ausgewählten Profilicons"
      petName:
        type: "string"
        example: "Sparky 2.0"
        description: "Name des PokeTier"
      petKind:
        type: "string"
        example: "KANGAROO"
        description: "Art des Poketier"
      petImage:
        type: "string"
        example: null
        description: "Optionales, in Base64 codiertes, Bild des PokeTier. Max. ca. 4 KB. \"null\" löscht das Bild."
...
  ApiErrorResponse:
    type: "object"
    properties:
      code:
        type: "string"
        example: "ERR_ACC_NOT_FOUND"
        description: "Ein Fehlercode, der bei der Programmierung des Clients helfen kann."
      devMsg:
        type: "string"
        example: "Account 'EpicRobloxMaster' wurde nicht gefunden."
        description: "Eine Nachricht, die Entwicklern helfen soll, den Fehler besser zu verstehen, aber nicht weiterverarbeitet werden soll."
      content:
        type: "object"
        example: null
        description: "Dieser Wert beinhaltet normalerweise die Daten der Antwort des Servers. Im Falle eines Fehlers ist dieser üblicherweise aus Kompatibilitätsgründen \"null\"."
  ApiSuccessResponseSimple:
    type: "object"
    properties:
      code:
        type: "string"
        example: "SUCCESS"
        description: "Ein Code, der zeigt, dass die Anfrage erfolgreich war. Dieser kann ignoriert werden, da man aus dem Statuscode schließen kann, dass es sich um einen Erfolg handelt."
      devMsg:
        type: "string"
        example: ""
        description: "Eine Nachricht, die Entwicklern zusätzliche Informationen geben kann, aber nicht weiterverarbeitet werden soll. Erfolgreiche Anfragen haben meistens keine Nachricht."
      content:
        type: "object"
        example: null
        description: "Dieser Wert beinhaltet die Daten der Antwort des Servers. Im Falle eines Fehlers ist dieser üblicherweise aus Kompatibilitätsgründen \"null\". Es kommt auf die Anfrage an, auf was dieser Wert gesetzt wird."
...